requirements.txt
click==8.1.7
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
tk==0.1.0
tkinterdnd2==0.3.0
numpy==1.26.4